need update:需要修改用户分页模块

+++ backend/src/main/java/com/hupi/project/model/vo/UserRoleVO.java
@@ -10,7 +10,7 @@ public class UserRoleVO {
 
     private User user;
     private Long[] roleIds;
-    private Integer[] deptIds;
+    private Long[] deptIds;
     private Long[] postIds;
 
 }
